move authentication handler and apollo client into components folder
